- Splitting Mixin's into individual files - Correcting Enum Name (ComparatorOptions)
