+++ src/main/java/carpetextra/mixins/ComparatorBlock_comparatorBetterItemFramesMixin.java
@@ -5,8 +5,6 @@ import net.minecraft.block.AbstractRedstoneGateBlock;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ComparatorBlock;
 import net.minecraft.entity.decoration.ItemFrameEntity;
-import net.minecraft.entity.player.PlayerAbilities;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.util.math.Direction;
@@ -14,7 +12,6 @@ import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.ArrayList;
@@ -22,22 +19,9 @@ import java.util.List;
 import java.util.Objects;
 
 @Mixin(ComparatorBlock.class)
-public abstract class ComparatorBlockMixin extends AbstractRedstoneGateBlock {
-    protected ComparatorBlockMixin(Settings settings) {
+public abstract class ComparatorBlock_comparatorBetterItemFramesMixin extends AbstractRedstoneGateBlock {
+    protected ComparatorBlock_comparatorBetterItemFramesMixin(Settings settings) {
         super(settings);
-    }
-
-    @Redirect(
-        method = "onUse",
-        at = @At(
-            value = "FIELD",
-            target = "Lnet/minecraft/entity/player/PlayerEntity;abilities:Lnet/minecraft/entity/player/PlayerAbilities;"
-        )
-    )
-    private PlayerAbilities hasPlayerAbilities(final PlayerEntity player) {
-        // player will never be null in VANILLA
-        if (player == null) return new PlayerAbilities();
-        return player.abilities;
     }
 
     private ItemFrameEntity getAttachedItemFrameHorizontal(World world, Direction facing, BlockPos pos) {
