+++ src/main/resources/carpet-extra.mixins.json
@@ -3,73 +3,71 @@
   "package": "carpetextra.mixins",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "DispenserBlockMixin",
-    "FallingBlockEntityMixin",
-    "RepeaterBlockMixin",
-    "ComparatorBlockMixin",
-    "DaylightDetectorBlockMixin",
-    "AnimalEntity_dispenserMixin",
-    "FoxMateGoalMixin",
-    "TurtleMateGoalMixin",
-    "AbstractPressurePlateBlock_variableWoodMixin",
-    "AbstractButtonBlock_variableWoodMixin",
-    "ItemFrameEntity_comparatorReadsClockMixin",
-    "ItemStackMixin",
-    "ThrownEntityMixin",
-    "BlockItemMixin_accurateBlockPlacement",
-    "HopperMinecartEntity_cooldownFixMixin",
-    "HopperMinecartEntity_transferItemsOutFeatureMixin",
     "AbsorbWaterMixin",
-    "DropperBlock_craftingMixin",
-    "DispenserBlockEntity_craftingMixin",
-
+    "AbstractButtonBlock_variableWoodMixin",
+    "AbstractPressurePlateBlock_variableWoodMixin",
+    "AbstractRedstoneGateBlockMixin",
+    "AnimalEntity_dispenserMixin",
+    "BlockItemMixin_accurateBlockPlacement",
+    "BlocksMixin",
+    "CactusBlock_fertilizerMixin",
+    "CactusBlock_syncMixin",
+    "ChickenEntityMixin",
+    "ChunkGeneratorMixin",
+    "ComparatorBlock_comparatorBetterItemFramesMixin",
+    "ComparatorBlock_playerAbilityMixin",
+    "DaylightDetectorBlockMixin",
+    "DispenserBehaviorBucketCowsMixin",
     "DispenserBehaviorCarpetsMixin",
+    "DispenserBehaviorCarvedPumpkinMixin",
     "DispenserBehaviorChestMixin",
     "DispenserBehaviorFireChargeMixin",
     "DispenserBehaviorGlowstoneMixin",
-    "DispenserBehaviorShearsMixin",
-    "DispenserBehaviorCarvedPumpkinMixin",
     "DispenserBehaviorMobHeadMixin",
-
-    "VillagerEntity_wartFarmMixin",
-    "FarmerVillagerTask_wartFarmMixin",
-    "VillagerTaskListProvider_wartFarmMixin",
-    "VillagerProfession_wartFarmMixin",
-    "SecondaryPointOfInterestSensor_wartFarmMixin",
-    "NoteBlockMixin",
-    "RedstoneWireBlockMixin",
-    "NetherWartBlock_fertilizerMixin",
-    "RedstoneWireBlockMixin_pistonRedirectsRedstone",
-
-    "PlayerEntityMixin",
-    "PistonBlockMixin",
-    "CactusBlock_syncMixin",
+    "DispenserBehaviorShearsMixin",
     "DispenserBlock_syncMixin",
-    "HopperBlock_syncMixin",
-    "SaplingBlock_syncMixin",
-    "SugarCaneBlock_syncMixin",
-    "BlocksMixin",
-    "ChunkGeneratorMixin",
-    "FallingBlockMixin",
-    "ChickenEntityMixin",
-    "LivingEntityMixin",
-    "FlowerPotBlockMixin",
+    "DispenserBlockEntity_craftingMixin",
+    "DispenserBlockMixin",
+    "DropperBlock_craftingMixin",
     "EntityMixin",
-    "DispenserBehaviorBucketCowsMixin",
-    "AbstractRedstoneGateBlockMixin",
+    "FallingBlockEntityMixin",
+    "FallingBlockMixin",
+    "FarmerVillagerTask_wartFarmMixin",
+    "FlowerPotBlockMixin",
+    "FoxMateGoalMixin",
+    "HopperBlock_syncMixin",
+    "HopperMinecartEntity_cooldownFixMixin",
+    "HopperMinecartEntity_transferItemsOutFeatureMixin",
     "IglooFeatureMixin",
-    "SurfaceChunkGeneratorMixin",
-    "StrayEntityMixin",
+    "ItemFrameEntity_comparatorBetterItemFramesMixin",
+    "ItemFrameEntity_comparatorReadsClockMixin",
+    "ItemStackMixin",
     "JungleTempleFeature_creeperMixin",
-    "SkeletonEntityMixin",
-    "SpiderEntityMixin",
+    "LilyPadBlock_fertilizerMixin",
+    "LivingEntityMixin",
+    "MinecraftServer_updateSuppressionCrashFixMixin",
+    "NetherWartBlock_fertilizerMixin",
+    "NoteBlockMixin",
+    "PistonBlockMixin",
+    "PlayerEntityMixin",
+    "RedstoneWireBlockMixin",
+    "RedstoneWireBlockMixin_pistonRedirectsRedstone",
+    "RepeaterBlockMixin",
+    "SaplingBlock_syncMixin",
     "ScaffoldingBlock_scaffoldingDistanceMixin",
     "ScaffoldingItem_scaffoldingDistanceMixin",
+    "SecondaryPointOfInterestSensor_wartFarmMixin",
+    "SkeletonEntityMixin",
+    "SpiderEntityMixin",
+    "StrayEntityMixin",
     "SugarCaneBlock_fertilizerMixin",
-    "CactusBlock_fertilizerMixin",
-    "LilyPadBlock_fertilizerMixin",
-
-    "MinecraftServer_updateSuppressionCrashFixMixin",
+    "SugarCaneBlock_syncMixin",
+    "SurfaceChunkGeneratorMixin",
+    "ThrownEntityMixin",
+    "TurtleMateGoalMixin",
+    "VillagerEntity_wartFarmMixin",
+    "VillagerProfession_wartFarmMixin",
+    "VillagerTaskListProvider_wartFarmMixin",
     "World_updateSuppressionCrashFixMixin"
   ],
   "injectors": {
